Move network definition (global variables) from header to c file to fix compile error without "extern" keyword.
The issue: we define the variables (with external linkage) in
our header.
When we include that header into multiple translation units,
i.e., nn_model & our additional "main" file, we define the
variables multiple times (with external linkage).
-> compile error.
In the arduino example it works as we "extern include" the
nn_model.h. However, that's not easily understandable and
not the way most people would include a header.